Restructured repo after NPM standards
