+++ package.json
@@ -2,7 +2,7 @@
   "name": "netlife-api",
   "version": "0.0.3",
   "description": "A simple to use package for Netlifes many API end-points.",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
     "test": "node tests/test.js"
   },
